CHANGED behaviour of "Smart Align (Left/Right/Top/Bottom)" to snap the selected layer to the next layer within the same group in that specific direction. Huh?
Example: "Smart Align Left" now moves (snaps) the selected layer to the right border of the next layer to the left in the same parent group. If there is no more layer to the left, it behaves as "Smart Align" did before, it moves the selected layer to the left border of the parent group or to the left border of the artboard.